fix(console): typecast operands to match data type
This corrects the MISRA violation C2012-10.3:
The value of an expression shall not be assigned to an object with a
narrower essential type or of a different essential type category.
Variable type has been updated from uint8_t to uint32_t to avoid
assigning a value to an object with a narrower essential type.

+++ drivers/console/multi_console.c
@@ -11,7 +11,7 @@
 #include <drivers/console.h>
 
 console_t *console_list;
-static uint8_t console_state = CONSOLE_FLAG_BOOT;
+static uint32_t console_state = CONSOLE_FLAG_BOOT;
 
 IMPORT_SYM(console_t *, __STACKS_START__, stacks_start)
 IMPORT_SYM(console_t *, __STACKS_END__, stacks_end)
